+++ src/azure-cli-core/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.0.28
 ++++++
 * Minor fixes
+* Expose telemetry to Azure CLI Extension
 
 2.0.27
 ++++++